requirements.txt
openai==1.25.0
python-dotenv==1.0.1
tiktoken==0.6.0
